[Modify] - 생성 페이지에 사용되는 컴포넌트 폴더 분리 - atom stores로 이동
