+++ src/components/index.ts
@@ -11,6 +11,6 @@ export { default as Footer } from '@/components/footer/Footer';
 export { default as Review } from '@/components/review/Review';
 export { default as Keyword } from '@/components/keyword/Keword';
 export { default as SwiperMain } from '@/components/swipers/SwiperMain';
-export { default as InputComponent } from '@/pages/create/components/InputComponent';
+export { default as InputComponent } from '@/components/input/InputComponent';
 export { default as RecipeCard } from '@/components/cards/recipeCard/RecipeCard';
 export { default as GlobalNavigationBar } from '@/components/navigation/GlobalNavigationBar';
