+++ src/pages/create/CreateThree.tsx
@@ -1,21 +1,23 @@
-import { InputComponent, Header, Button } from "@/components"
-import TextAreaComponent from "./components/TextAreaComponent";
+import { Header, Button } from "@/components"
+import { TextAreaComponent, FileInputComponent } from "./components/";
 import { useAtom } from "jotai";
 import { useState } from 'react';
-import { recipeSteps } from ".";
+import { recipeSteps } from '@/stores/stores';
 import { Link } from "react-router-dom";
+import { getRandomId } from "@/util/math/getRandomId";
 
 
 export function CreateThree() {
   const [steps, setSteps] = useAtom(recipeSteps);
   const [description, setDescription] = useState('');
   const [tips, setTips] = useState('');
+  const [image, setImage] = useState<File | null>(null);
 
   return (
     <div className="flex flex-col h-full">
       <Header option="titleWithClose" title="레시피 스탭 추가하기" />
       <div className="flex flex-col px-16pxr py-14pxr grow w-full gap-42pxr">
-        <InputComponent option="fileInput" inputTitle="이미지" />
+        <FileInputComponent inputTitle="단계 이미지" setFile={setImage}/>
         <TextAreaComponent 
           inputTitle="설명" 
           requiredText=" (필수)" 
@@ -37,7 +39,8 @@ export function CreateThree() {
           onClick={() => {
             console.log(description, tips);
             // change image value to the image url
-            setSteps(JSON.stringify([...JSON.parse(steps), {"image": '', "description": description, "tips": tips}])) 
+            const id = getRandomId();
+            setSteps(JSON.stringify([...JSON.parse(steps), {"id":id,"image": image, "description": description, "tips": tips}])) 
             console.log(steps)
           }}>완료
         </Link>
@@ -45,3 +48,4 @@ export function CreateThree() {
     </div>
   )
 }
+
