+++ src/pages/create/components/TextAreaComponent.tsx
@@ -8,7 +8,7 @@ interface TextAreaProps {
   setData : Dispatch<SetStateAction<string>>;
 }
 
-export default function TextAreaComponent({ inputTitle, requiredText, maxCharCount, setData  }: TextAreaProps) {
+export function TextAreaComponent({ inputTitle, requiredText, maxCharCount, setData  }: TextAreaProps) {
   const [text, setText] = useState('');
   const [charCount, setCharCount] = useState(0);
