+++ src/pages/create/components/index.ts
@@ -1,0 +1,7 @@
+
+
+
+export * from './FileInputComponent.tsx'
+export * from './IngredientsComponent.tsx'
+export * from './TextAreaComponent.tsx'
+export * from './TitleComponent.tsx'
