+++ src/pages/user/MyRecipes.tsx
@@ -1,5 +1,5 @@
 import { useAtom } from "jotai"
-import { ingredients } from ".."
+import { ingredients } from "@/stores/stores"
 
 export function MyRecipes() {
   const [ingredientData, setIngredientData] = useAtom(ingredients);
